Separate customize mode from the `export` subcommand
